new uvs2 releases : uppsrc-2598  tutorial-38  examples-141  reference-113
git-svn-id: svn://ultimatepp.org/upp/trunk@300 f0d560ea-af0d-0410-9eb7-867de7ffcac7

+++ uppsrc/Core/Stream.cpp
@@ -274,19 +274,19 @@ String Stream::GetLine() {
 	byte *q = ptr;
 	while(q < rdlim)
 		if(*q == '\n') {
-			String result((const char *)ptr, q - ptr - (q > ptr && q[-1] == '\r'));
+			String result((const char *)ptr, (int)(uintptr_t)(q - ptr - (q > ptr && q[-1] == '\r')));
 			ptr = q + 1;
 			return result;
 		}
 		else
 			q++;
-	String result((const char *)ptr, q - ptr);
+	String result((const char *)ptr, (int)(uintptr_t)(q - ptr));
 	ptr = q;
 	for(;;) {
 		byte *q = ptr;
 		while(q < rdlim && *q != '\n')
 			q++;
-		result.Cat(ptr, q - ptr);
+		result.Cat(ptr, (int)(uintptr_t)(q - ptr));
 		ptr = q;
 		int c = Get();
 		if(c == '\n')
